Build UserInput component skeleton.
Created the components folder and inside it a new "UserInput.jsx" file
in which a hardcoded UserInput component was built according to the CSS
requirements. This component contains a form with two different groups
of two Label/Input pair each. These inputs will be used by the user to
enter the data to make the investment calculations.