Polish UserInputSection component
UserInput component was renamed to UserInputSection.

Now instead of having a hardcoded UserInputSection component, it is
given dynamism by creating the Label/Input pairs as a sub-component in
another file.

This sub-component is named Input and doesn't use hardcoded values.
Instead, it uses props to set up the different attributes that the
elements need.

The values for the Input component attributes come from the
"INPUTS_DATA" variable, which is also located in a separate file. Those
values are automatically spread within the required attributes according
to the index of the object they come from.