Add fetching data functionality.
Built the fetching data functionality within the UserInputSection
component, so that the "useState" update only affects this component and
not the whole App.

Had to update "INPUTS_DATA" variable and "Input" component in order to
add a new "className" prop which contains the identifiers which are used
by the "calculateInvestmentResults" function to transform the user's
data into the investment results.

Fetch of data is accomplished through an "onChange" prop built in the
Input elements which triggers the "handleInputChange" function. This
function triggers the state update and updates an object which contains
the necessary key/value pairs for each piece of data but empty. The
object is looped and every key name compared to the className of the
Input element that triggered the function. The matching Input element'so
value updates the matching key from the object, thus storing the data.