Fix "handleInputChange", "investmentData", and create "investmentResults".
Variable "investmentResults" is created within the "ResultsTable"
component to store the results array created after applying function
"calculateInvestmentResults" to the "userData" object.

In order for the function "calculateInvestmentResults" to transform the
data properly, a few changes were needed:
--> const "investmentData" values were changed to null.
--> values are transformed to numbers before updating the object keys.

An if... else statement was added after checking equality between the
object key and the input className, so that if the className corresponds
to the "duration" input and its current value is below "0", an alert
will appear asking the user that the input needs a positive value, and
its current value is replaced by "null".