Build array storing the final results to be displayed.
Function "buildResultsArray" is created outside the "ResultsTable"
component so it can be used inside it, while keeping the component clean.
It takes the "userData" and "investmentResults" variables as parameters
to create an Array of subarrays, which contains all the necessary data
to be displayed, divided by rows according to the years of the investment
duration and ordered to match the headers.

Then, variable "finalResultsArray" is created within the "ResultsTable"
component by executing the "buildResultsArray" function with its
necessary parameters. This variable will be mapped to display the results
within the component's table.