Finish results display functionality within "ResultsTable" component.
The results part of the table is built by mapping through the
"finalResultsArray" array and its sub-arrays and displaying all its
values. The first value of each sub-array is displayed as it comes but
the rest of them are passed first through the "formatter" function in
order to give them the correct display format (a round up to become
integers and a dollar currency status).